samples/start-app/python-fastapi/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0 
requests>=2.28.0,<=2.29.0
python-dotenv>=0.21.1,<=0.22.0
sqlalchemy>=2.0.0,<=2.0.1
